Setup controller, service, dto, and swagger
