+++ src/app.controller.ts
@@ -1,5 +1,6 @@
-import { Controller, Get, Param } from '@nestjs/common';
-import { AppService, FindUserDto, CreateUserDto } from './app.service';
+import { Controller, Get, Post, Put, Delete, Param, Body } from '@nestjs/common';
+import { AppService } from './app.service';
+import { FindUserDto, CreateUserDto } from './user.dto';
 
 @Controller()
 export class AppController {
@@ -18,5 +19,20 @@ export class AppController {
 	@Get(':id')
 	getUser(@Param() params: FindUserDto): string {
 		return this.appService.getUser(params.id);
+	}
+	
+	@Post(':id')
+	updateUser(@Param() params: FindUserDto): string {
+		return this.appService.updateUser(params.id);
+	}
+	
+	@Delete(':id')
+	deleteUser(@Param() params: FindUserDto): string {
+		return this.appService.deleteUser(params.id);
+	}
+	
+	@Put()
+	addUser(@Body() params: CreateUserDto): string {
+		return this.appService.addUser(params);
 	}
 }
